# apis/app-api/order/order-app-api.openapi.json
{
  "openapi": "3.1.2",
  "info": {
    "title": "SDKWork Order App API",
    "version": "1.0.0",
    "description": "App/client contract for sdkwork-order capability (split from retired sdkwork-commerce monolith).",
    "x-sdkwork-api-authority": "sdkwork-order-app-api",
    "x-sdkwork-sdk-family": "sdkwork-order-app-sdk",
    "x-sdkwork-audience": "App, desktop, mobile, H5, and user-facing clients.",
    "x-sdkwork-owner": "sdkwork-order"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:18079",
      "description": "Local sdkwork-order via IM gateway"
    }
  ],
...
  "paths": {
    "/app/v3/api/checkout/sessions": {
      "post": {
        "tags": [
          "orderCheckout"
        ],
        "summary": "Checkout sessions create.",
        "operationId": "checkout.sessions.create",
        "parameters": [
          {
            "$ref": "#/components/parameters/WriteCommandIdempotencyKey"
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCheckoutSessionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-order",
        "x-sdkwork-api-authority": "sdkwork-order-app-api",
        "x-sdkwork-domain": "commerce",
        "x-sdkwork-resource": "checkout.sessions",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCheckoutSessionRequest"
              }
            }
          }
        },
        "x-sdkwork-idempotent": true
      }
    },
    "/app/v3/api/checkout/sessions/{checkoutSessionId}": {
      "get": {
        "tags": [
          "orderCheckout"
        ],
        "summary": "Checkout sessions retrieve.",
        "operationId": "checkout.sessions.retrieve",
        "parameters": [
          {
            "name": "checkoutSessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutSessionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-order",
        "x-sdkwork-api-authority": "sdkwork-order-app-api",
        "x-sdkwork-domain": "commerce",
        "x-sdkwork-resource": "checkout.sessions",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api"
      }
    },
    "/app/v3/api/checkout/sessions/{checkoutSessionId}/quotes": {
      "post": {
        "tags": [
          "orderCheckout"
        ],
        "summary": "Checkout sessions quotes create.",
        "operationId": "checkout.sessions.quotes.create",
        "parameters": [
          {
            "name": "checkoutSessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/WriteCommandIdempotencyKey"
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCheckoutQuoteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-order",
        "x-sdkwork-api-authority": "sdkwork-order-app-api",
        "x-sdkwork-domain": "commerce",
        "x-sdkwork-resource": "checkout.sessions.quotes",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api",
        "x-sdkwork-idempotent": true
      }
    },
    "/app/v3/api/checkout/sessions/{checkoutSessionId}/orders": {
      "post": {
        "tags": [
          "orderCheckout"
        ],
        "summary": "Checkout sessions orders create.",
        "operationId": "checkout.sessions.orders.create",
        "parameters": [
          {
            "name": "checkoutSessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/WriteCommandIdempotencyKey"
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCheckoutOrderResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-order",
        "x-sdkwork-api-authority": "sdkwork-order-app-api",
        "x-sdkwork-domain": "commerce",
        "x-sdkwork-resource": "checkout.sessions.orders",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api",
        "x-sdkwork-idempotent": true
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AuthToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "SDKWork auth token carried as Authorization: Bearer <auth_token>."
      },
      "AccessToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Access-Token",
        "description": "SDKWork access isolation token."
      }
    },
    "schemas": {
...
      "CreateCheckoutSessionRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "minItems": 1,
            "items": {
              "$ref": "#/components/schemas/CheckoutLineRequest"
            }
          },
          "currencyCode": {
            "type": "string",
            "minLength": 3,
            "maxLength": 3,
            "default": "CNY"
          },
          "shippingAddress": {
            "$ref": "#/components/schemas/ShippingAddressRequest",
            "description": "实物商品必填；数字商品（如兑换码、虚拟权益）可省略。"
          }
        }
      },
...
      "CreateCheckoutSessionResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SdkWorkApiResponse"
          },
          {
            "type": "object",
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "type": "object",
                "required": [
                  "item"
                ],
                "properties": {
                  "item": {
                    "$ref": "#/components/schemas/CheckoutSession"
                  }
                }
              }
            }
          }
        ]
      },
      "CheckoutSessionResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SdkWorkApiResponse"
          },
          {
            "type": "object",
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "type": "object",
                "required": [
                  "item"
                ],
                "properties": {
                  "item": {
                    "$ref": "#/components/schemas/CheckoutSession"
                  }
                }
              }
            }
          }
        ]
      },
...
      "ProblemDetail": {
        "type": "object",
        "additionalProperties": true,
        "required": [
          "type",
          "title",
          "status",
          "code",
          "traceId"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri-reference"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": 100,
            "maximum": 599
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "code": {
            "$ref": "#/components/schemas/SdkWorkPlatformErrorCode"
          },
          "traceId": {
            "type": "string",
            "format": "uuid",
            "description": "Server-owned request correlation id."
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          }
        }
      },
...
    },
    "parameters": {
      "WriteCommandIdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 128
        },
        "description": "Client retry key scoped by tenant, principal, method, and path."
      }
    }
  },
  "security": [
    {
      "AuthToken": [],
      "AccessToken": []
    }
  ]
}